+++ drivers/frame_provider/decoder/vc1/vvc1.c
@@ -516,7 +516,7 @@ static irqreturn_t vvc1_isr(int irq, void *dev_id)
 					}
 					if (next_pts_us64 != 0) {
 						next_pts_us64 +=
-						((vf->duration) -
+						(u64)((vf->duration) -
 						((vf->duration) >> 4)) *
 						100 / 9;
 					}
@@ -582,7 +582,7 @@ static irqreturn_t vvc1_isr(int irq, void *dev_id)
 						 ((vf->duration) >> 4));
 				}
 				if (next_pts_us64 != 0) {
-					next_pts_us64 += ((vf->duration) -
+					next_pts_us64 += (u64)((vf->duration) -
 					((vf->duration) >> 4)) * 100 / 9;
 				}
 			} else {
@@ -673,7 +673,7 @@ static irqreturn_t vvc1_isr(int irq, void *dev_id)
 					}
 					if (next_pts_us64 != 0) {
 						next_pts_us64 +=
-						((vf->duration) -
+						(u64)((vf->duration) -
 						((vf->duration) >> 4)) *
 						100 / 9;
 					}
